pyproject.toml
[tool.poetry]
name = "fast-api-backend"
version = "0.1.0"
description = "FastAPI backend template"
authors = ["ritesh-dubey <ritesh.dubey@wowlabz.com>"]
readme = "README.md"
packages = [{include = "fast_api_backend"}]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
pre-commit = "3.3.3"
aiofiles = "23.1.0"
aioredis = "2.0.1"
aiosmtplib = "2.0.1"
azure-storage-blob = "12.16.0"
boto3 = "1.26.138"
fastapi = {version = "0.95.2", extras = ["all"]}
firebase-admin = "6.1.0"
httpx = "0.24.1"
loguru = "0.7.0"
motor = "3.1.2"
opencv-python-headless = "4.7.0.68"
orjson = "3.8.12"
pandas = "1.4.3"
pip-licenses = "4.3.2"
pybars3 = "0.9.7"
python-jose = {version = "3.3.0", extras = ["cryptography"]}
python-multipart = "0.0.6"
sendgrid = "6.10.0"
twilio = "8.5.0"
user-agents = "2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
